Message signing & cryptography package (#21)
Sign messages from gateway to client.
Crypto package added to the client.
Load private key from settings.

+++ internal/reputation/reputation_constants.go
@@ -2,40 +2,40 @@ package reputation
 // Copyright (C) 2020 ConsenSys Software Inc
 
 
-const clientMaxReputation = 10000
-const clientMinReputaiton = -10000
-const clientInitialReputation = 10
-const clientEstablishmentChallenge = -1
-const clientOnChainDeposit = 1000
+const clientMaxReputation = int64(10000)
+const clientMinReputaiton = int64(-10000)
+const clientInitialReputation = int64(10)
+const clientEstablishmentChallenge = int64(-1)
+const clientOnChainDeposit = int64(1000)
 
 //Response with one or more CID Offers. Initial payment and final payment made.
-const clientStdDiscOneCidOffer = 10
+const clientStdDiscOneCidOffer = int64(10)
 
 // Response with no CID Offers. Initial payment payment made.
-const clientStdDiscNoCidOffers = 1
+const clientStdDiscNoCidOffers = int64(1)
 
 // Response with one or more CID Offers. Response message sent after one second prior to TTL expiry. Initial payment payment made.
-const clientStdDiscLateCidOffers = 1
+const clientStdDiscLateCidOffers = int64(1)
 
 // Response with one or more CID Offers. Response message sent prior to one second prior to TTL expiry. Initial payment payment made but final payment not paid.
-const clientStdDiscNonPayment = -100
+const clientStdDiscNonPayment = int64(-100)
 
 // Response with one or more CID Offers from one or more Gateways. Initial payment and final payment made
-const clientDhtDiscOneCidOffer = 10
+const clientDhtDiscOneCidOffer = int64(10)
 
 // Response with no CID Offers. Initial payment made.
-const clientDhtDiscNoCidOffers = 1
+const clientDhtDiscNoCidOffers = int64(1)
 
 // Response with one or more CID Offers. Response message sent after one second prior to TTL expiry. Initial payment payment made.
-const clientDhtDiscLateCidOffers = 1
+const clientDhtDiscLateCidOffers = int64(1)
 
 // Response with one or more CID Offers. Response message sent prior to one second prior to TTL expiry. Initial payment payment made but final payments not paid.
-const clientDhtDiscNonPayment = -300
+const clientDhtDiscNonPayment = int64(-300)
 
 // Micro-payment paid for content via Gateway. Note that there will be many micro-payments during content retrieval.
-const clientMicroPayment = 1
+const clientMicroPayment = int64(1)
 
 // Invalid message received
-const clientInvalidMessage = -10
+const clientInvalidMessage = int64(-10)
